camara-api/requirements.txt
streamlit==1.52.1
requests==2.32.5
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.0.0
beautifulsoup4==4.12.3
playwright==1.58.0
